Cmake generated project. Switches to x64 as a side effect. Had to update glew and make some size fixes. Still missing tests and models aren't loading for some reason.
Add support for tests in cmake project

Add precompiled headers and make headers show up properly in intellisense

Warnings as errors, disable exceptions, fix warnings.

Fix incorrect element size causing garbage vertex buffers and restore scene.json

Fix warning, set default startup project

Remove visual studio related files

Add visual studio files to gitignore

+++ syx/Engine.Shared/ImGuiImpl.cpp
@@ -223,7 +223,7 @@ void ImGuiImpl::render(float dt, Syx::Vec2 display) {
 
       for(int cmd = 0; cmd < cmdList->CmdBuffer.Size; ++cmd) {
         const ImDrawCmd* pcmd = &cmdList->CmdBuffer[cmd];
-        GLuint curTex = reinterpret_cast<GLuint>(pcmd->TextureId);
+        GLuint curTex = static_cast<GLuint>(reinterpret_cast<size_t>(pcmd->TextureId));
 
         glScissor((int)pcmd->ClipRect.x, (int)(fbHeight - pcmd->ClipRect.w), (int)(pcmd->ClipRect.z - pcmd->ClipRect.x), (int)(pcmd->ClipRect.w - pcmd->ClipRect.y));
         if(curTex != boundTex) {
